codewhynot2/src/causal/scm_builder.py
import networkx as nx
import yaml

class SCMBuilder:
    def __init__(self, config_path: str):
        self.config_path = config_path

    def build_graph(self):
        """Build a DAG from a YAML/JSON SCM definition."""
        # Implementation placeholder
        return nx.DiGraph()
